fix(ci): split dot convert to svg into separate step (#1546)
### Briefly, what does this PR introduce?
Actions fail on dot with `Error: trouble in init_rank`. E.g.
https://github.com/eic/EICrecon/actions/runs/10234743256/job/28315055948.
Since this error occurs in the EICrecon step, it may be incorrectly
interpreted as an EICrecon failure.

This PR splits the step in two. It is the first step in figuring out
what's wrong with dot. Turns out we throw away the source and sink info
that helps with rank determination. This PR also puts the source and
sink back (and only removes `podio::Frame` and JEventProcessorPODIO`
from them).

### What kind of change does this PR introduce?
- [x] Bug fix (issue: trouble in init_rank)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.