fix: avoid division by zero in UndoAfterBurner when no beam particles (#1553)
### Briefly, what does this PR introduce?
When no beam particles are in the MCParticles event record, the
UndoAfterBurner performs a division by zero when calculating the boost
beta vector since both incoming electron and hadron momentum are zero
(so their sum is). See e.g.
https://github.com/eic/EICrecon/actions/runs/10239218932/job/28324710737#step:9:6197.

This PR causes the algorithm to bail out when it is unable to find (or
assume) an incoming electron and hadron beam particle.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: division by zero)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @ajentsch
@bspage912

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.