lowered the amount of distance neighbor hits in topoclustering in the ZDC (#1532)
### Briefly, what does this PR introduce?
Redefines neighboring subcell hits in the ZDC (in the context of topo
clustering) to be the nearest neighbors only for hits in the same layer,
and the same position for hits in adjacent layers. This change reduces
the cases where showers from different particles are merged into a
single cluster

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: parameter tweak

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no.  
### Does this PR change default behavior?
yes.  Changes the default behavior of the topo clustering in the ZDC

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Derek M Anderson <derek.murphy.anderson@protonmail.com>