fix: deterministic cluster time (#1584)
### Briefly, what does this PR introduce?
This PR makes the cluster time deterministic, and not based on the whim
of which hit happens to be first. It defines the cluster time as the
energy-weighted average of all hit times.

This is not the only deterministic choice here (earliest hit is
another).

### What kind of change does this PR introduce?
- [x] Bug fix (issue: cluster time depends on hit ordering)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.