fix(ci): use {ref,new} labels on capybara plots (#1602)
### Briefly, what does this PR introduce?
Minor PR that change the labels on the capybara plots from `ref` and
`rec_dis_18x275_minQ2=0_craterlake.edm4eic.root`, to `ref` and `new`.
The shorter names are clearer, shorter, and don't push the legend's line
type off the plotting canvas (which makes it really useless). The name
of the file is in the URL as well.

Now looks like:

![image](https://github.com/user-attachments/assets/c8c55f9a-4110-499f-b987-05e76503ce04)

### What kind of change does this PR introduce?
- [x] Bug fix (issue: sometimes capybara legend line type is pushed off
the plotting canvas)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.