allow MultiSegmentation in calorimeter hit recon (#1594)
Allow multisegmentation without giving a warning about it.

### Briefly, what does this PR introduce?
There is an if statement in CalorimeterHitReco that checks if the
segmentation type is one of several types of segmentations (cartesian or
hexagonal) that prints a warning if the segmentation is not one of
those. Adds MultiSegmentation to this list.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: removal of unnecessary warning

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no
### Does this PR change default behavior?
no

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nathan Brei <nbrei@jlab.org>