feat(ci): store prmon summary as artifact (#1603)
### Briefly, what does this PR introduce?
This PR adds prmon monitoring of the reconstruction jobs eicrecon-gun
and eicrecon-dis, and uploads the summary as an artifact. This can be
used later as a way to fail when execution time or memory usage is
significantly changed by a PR. We first need to get a sense of how much
variability there is due to the GitHub backend nodes.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: catch large slowdowns in performance sooner)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.