HcalEndcapPInsertImagingProtoClusters: use a larger hit spacing threshold (#1595)
### Briefly, what does this PR introduce?
Uses a larger threshold for the distance between the neighboring hits in
the insert, due to the new cell size. Also include "side" field (ie,
left vs right) in the hit merger algorithm

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: change of parameters

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Probably. Including the detector "side" as a field in the hit merger
algorithm probably breaks the code if using a version of the epic
repository from before pull request https://github.com/eic/epic/pull/771
is merged. Therefore this pull request and
https://github.com/eic/epic/pull/771 should be merged at the same time
as one another
### Does this PR change default behavior?
yes. Changes the parameters of the topo clustering and also uses the
"side" field in the hit merger.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>