tracking: use Seeded tracks by default, rename previous default to TruthSeeded (#1605)
### Briefly, what does this PR introduce?

As we agreed during the collaboration meeting, the real seeding workflow
is mature and should become the default. This changes Seeded collections
to not include that prefix, and prefixless collections are moved to
TruthSeeding (name is long but not supposed to be used by general
users). As a side effect, Jet finding and ReconstructedParticles and
DRICH IRT-based PID will use real seeding.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Yes