updated MIP value in insert; also adjusted parameters of the topo clu… (#1611)
…stering

### Briefly, what does this PR introduce?
Updates the MIP value used in the HEXPLIT algorithm in the calorimeter
insert. Also lowers the threshold for the center hit in a cluster. Doing
this fixes issues in the benchmark for neutron reconstruction ( see
https://github.com/eic/physics_benchmarks/issues/32 ).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: Parameter update

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no

### Does this PR change default behavior?
yes.  Changes the parameters of the topo clustering.