Correct xml file being called for different beam settings in CI (#1616)
### Briefly, what does this PR introduce?
Fixes xml being used so that it matches the MC sample being run.

It was noted in the Exclusive Diffractive and Tagging meeting that some
of the campaigns appear to be being run with the wrong beamline
settings. This only address the problem where I can see it in the CI.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Means the beam is being steered correctly for the MC samples being run
in the CI

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>