Add Track-Based Cluster Merging Algorithm (#1406)
### Briefly, what does this PR introduce?

This PR introduces the `TrackClusterMergeSplitter` algorithm, an
algorithm for merging and splitting calorimeter clusters based on
matched track projections. Inspired by the [Particle Flow algorithm from
ATLAS](https://doi.org/10.1140/epjc/s10052-017-5031-2).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #897 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No.

### Does this PR change default behavior?

Yes. Now "SplitMerge" `edm4eic::Cluster` collections and corresponding
associations will be created for 6 of the central calorimeters (`NHCal`,
`EEEMCal`, `BHCal`, `BIC`, `LFHCal`, and `FEMC`). Track projections will
need to be extended to other calorimeters before the algorithm
introduced here can be used on them.

### To-Do

The split-merge parameters for each calorimeter need to be updated from
the current dummy values used for testing. In the interest of getting
the functionality of this algorithm merged, this could update could be
done in a separate PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>