Create Multiple Primary MC-Cluster Associations (#1396)
### Briefly, what does this PR introduce?

This PR adjusts the logic used in establishing the truth-cluster
associations in `CalorimeterClusterRecoCoG`. Previously, the truth in
the truth-cluster associations was defined to be the MCParticle of the
1st contribution stored in the list of contributions to the
SimCalorimeterHit corresponding to the highest energy hit in the
cluster. This PR changes the logic such that now

1. Multiple MC-Cluster associations are created,
2. Associations link back to primary (`generator status == 1`) MC
particles, and
3. Each association carries a weight of `E_{contributed} / E_{cluster}`.

This PR addresses #1475 and partially addresses #898. The `generator
status == 1` condition should be revisited in a subsequent PR: this
definition of primary may miss cases, particularly for longer lived
neutral resonances such as the `K^{0}_{L}`, which would be useful
information to retain.

### What kind of change does this PR introduce?a
- [X] Bug fix (issues #898, #1475)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

While this PR doesn't introduce breaking changes, it does introduce
substantial changes that users might need to account for in downstream
analysis code (see below).

### Does this PR change default behavior?

**Yes:** MC-Cluster associations
(`edm4eic::MCRecoClusterParticleAssociations`) will now contain
_multiple_ associations per clusters, and will link back to primary
(`status == 1`) MC particles.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>