increased w0 parameter in cluster CoG for the ZDC (#1617)
increases the w0 parameter in the ZDC cluster reconstruction CoG so that
the subcell with the largest energy does not dominate the position
reconstruction too much. Without this fix, the positions of the clusters
are strongly pulled towards the subcell with the most energy, causing an
unwanted pattern in the reconstructed cluster position.

### Briefly, what does this PR introduce?
Increase the w0 parameter in the ZDC

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: parameter tweaking

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no
### Does this PR change default behavior?
yes.  modifies the paramters of the cluster position reconstruction