fix(ci): avoid python3.10 reference; don't install bokeh, pygithub (#1629)
### Briefly, what does this PR introduce?
We are using python3.11 now, so the `pip install` would install to
`~/.local/lib/python3.11/site-packages` if it did anything, and the
hardcoded `PYTHONPATH` prepend would then fail to find the module. That
this continued to work demonstrates that we don't need to install
anything locally, and can get rid of the `PYTHONPATH` modifications too.

This PR removes the `PYTHONPATH` additions and the custom `pip install`.
If an update to the bokeh and pygithub packages are needed, then this is
better handled by updating what is in the container, now that we have
reached a more stable capybara stage of development.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: hardcoded reference to python3.10)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.