fix(ci): always run capybara, even when ref is not found (#1630)
### Briefly, what does this PR introduce?
When, for some reason, the reference artifacts on `main` are not
available (typically recovering from a failing `main`), then we don't
generate capybara output and the `build-docs` job fails when trying to
merge the capy files. E.g.
https://github.com/eic/EICrecon/actions/runs/11282255676/job/31379983435.
There is no gracious "don't fail when you can't merge anything" option.

This PR ensures that we always have capy files generated, even when we
don't actually have a reference to compare to. The only thing that will
be included if there are is reference is the new file. That's still
useful information to plot.

### What kind of change does this PR introduce?
- [x] Bug fix (issue
https://github.com/eic/EICrecon/actions/runs/11282255676/job/31379983435)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.