Fix RP reco bug in p_y (#1634)
### Briefly, what does this PR introduce?

This PR introduces a bug fix for the p_y reconstruction in the Roman
pots, and also switches the reco to global coordinates. There are some
issues with the basic RP concept and how local coordinates are
calculated using the cellIDs. It means an additional set of shifts would
need to be calculated to do a proper transformation to the coordinate
system local to the proton orbit. Using global coordinates simplifies
this to the same procedure as before, but with different offsets (the
matrices themselves don't actually change).

### What kind of change does this PR introduce?
- [ xx] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x ] Tests for the changes have been added
- [x] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

See attached slides for some discussion and comparison plots.

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, it fixes the default momentum reconstruction to be correct (within
the limits of how the static matrix approach works). Users will only
need to change the axis limits on their 2D plots of the hits on the
Roman pots (see the Roman pots geometry definition to pick a
more-reasonable range).

### Does this PR change default behavior?

Yes, it improves the pT reconstruction for the Roman pots protons.

[ePIC_roman_pots_reco_matrix_bug_fix_10_15_2024.pdf](https://github.com/user-attachments/files/17382021/ePIC_roman_pots_reco_matrix_bug_fix_10_15_2024.pdf)

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>