Create Multiple Primary MC-Cluster Associations for Topoclusters (#1631)
### Briefly, what does this PR introduce?

This PR applies the #1396 treatment to the `ImagingClusterReco`
algorithm. As a reminder:
- Previously, the truth in the truth-cluster associations was defined to
be the MCParticle of the 1st contribution stored in the list of
contributions to the SimCalorimeterHit corresponding to the highest
energy hit in the cluster.
- The logic here is
    1. Multiple MC-Cluster associations are created,
2. Associations link back to primary (generator status == 1) MC
particles, and
3. Each association carries a weight of E_{contributed} / E_{cluster}.

### What kind of change does this PR introduce?
- [x] Bug fix (issues #898, #1475)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

It doesn't introduce breaking changes, but it does substantially alter
cluster-MC associations in such a way that users may have account for in
downstream analysis code.

### Does this PR change default behavior?

**Yes:** MC-Cluster associations
(edm4eic::MCRecoClusterParticleAssociations) will now contain multiple
associations per clusters, and will link back to primary (status == 1)
MC particles.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Chao Peng <pengc2010@gmail.com>