feat(ci): runs-on: ubuntu-latest -> ubuntu-22.04 (#1687)
### Briefly, what does this PR introduce?
This PR updates the target operating system for CI jobs, from
ubuntu-latest to a more explicit ubuntu-22.04. This is in part to avoid
the automatic update of implicit ubuntu-latest that will break
everything (https://github.com/eic/EICrecon/pull/1686), and in part
because it makes sense to be explicit about the target operating system.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue
https://github.com/actions/runner-images/issues/10636)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.