Allow xrootd files as inputs to eicrecon (#1690)
### Briefly, what does this PR introduce?
Currently eicrecon has two checks for files existing, neither of which
allow the file to be on a remote xrootd server. This PR adapts
JEventSourcePODIO.cc so that the first check allows for remote files and
(currently) removes the second check.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
tbd

### Does this PR change default behavior?
Allows remote files to be passed directly to eicrecon

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>