feat(ci): runs-on: ubuntu-22.04 -> ubuntu-24.04 (#1686)
### Briefly, what does this PR introduce?
This PR updates the target operating system for CI jobs, from
ubuntu-latest to a more explicit ubuntu-24.04. This is in part to test
that the automatic update of implicit ubuntu-latest will not break
anything (https://github.com/actions/runner-images/issues/10636), and in
part because it makes sense to be explicit about the target operating
system.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue
https://github.com/actions/runner-images/issues/10636)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, for the next few days ubuntu-latest will still be ubuntu-22.04, and
this will mean an upgrade to ubuntu-24.04.