Low-Q2 particle reconstruction using onnx framework (#1697)
### Briefly, what does this PR introduce?
Adds ML workflow for the Low-Q2 Tagger particle reconstruction based on
the onnx framework provided. As an alternative to the TMVA description.

Currently not added by default as there are a number of things which
need to be adapted. I will follow up with a note of the relevant
issues/PRs.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>