feat: support EDM4hep-0.99 relations to MCParticle (#1714)
### Briefly, what does this PR introduce?
In EDM4hep v0.99,
https://github.com/key4hep/EDM4hep/commit/cb738b0e2a5a8d4c922aa2a0d49974caf4adcd2e,
the names of relations to `edm4hep::MCParticle` types were renamed from
`MCParticle` to `particle`. This is more consistent in capitalization,
and it avoids reusing type names as field names. It also is necessary
for julia ports, it seems.

This PR adapts to the new names and avoids deprecation warnings during
compilation of EICrecon..

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: EDM4hep 0.99 support without warnings)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>