Correct PID LUT for protons (#1740)
### Briefly, what does this PR introduce?
The electron is checked for twice rather than including the proton
probability for an entry. Meaning anything with 100% chance of being a
proton is ignored.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Will set particles which have 100% chance of being identified as a
proton to have a proton PDG value where currently they will be set to 0.