Realistic BTOF digitization (#1635)
### Briefly, what does this PR introduce?

It's an attempt to make LGAD sensor response for Barrel TOF more
realistic by doing the following,

1. Spread charge deposition from strips that are hit directly to nearby
strips.
2. Perform digitization by converting charge deposited to electric
pulse.
3. Digitize the pulse by converting it to TDC and ADC value. ADC propto
pulse height and TDC propto time when voltage crosses certain threshold.
4. Convert the TDC and ADC value back to time and Edep by linear fit.
5. Return the BTOF hit point as "TOFBarrelRecHit". Position is estimated
by either weighted sum of ADC or just center of the cell with max ADC.
It's weighted average by default, but the behavior can be changed by
parameters.

Noise, edge correction and time talk correction will be developed in the
future.

### What kind of change does this PR introduce?
- [x ] Bug fix (issue #1656 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x ] Other: Implementation of digitization and clusterization for
LGADs in Barrel TOF

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

It replaces the content of "TOFBarrelRecHit" with results of this new
clusterization. It was originally just geant point + random smearing. I
have communicated to the reconstruction group and they advise that I
submit this pull request so they can study the effect of having this new
digitization/clusterization algorithm. They will decide if I should save
the data in a new branch or replace the origin content.

---------

Co-authored-by: Souvik Paul <136603463+eicsouvik@users.noreply.github.com>
Co-authored-by: Desktop ssedd1123 <ssedd1123@DESKTOP-0J5IETB>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Derek M Anderson <derek.murphy.anderson@protonmail.com>
Co-authored-by: Simon Gardner <simon.gardner@glasgow.ac.uk>